refactor(auth,trash,files): 重构刷新令牌、清理服务及文件预览逻辑
- 将 `refresh_tokens` 逻辑抽取至独立子模块 `tokens/refresh.rs`，使用 `tokio::Mutex` 替代简单占位锁，支持并发竞争等待而非直接拒绝，并新增 `RotateConflict` 错误分支处理旋转冲突后的重分类
- 修复刷新令牌旋转锁丢失时的分类逻辑，区分"近期同客户端"（StaleRefresh）与"真实重用"（ReuseDetected）场景，补充边界条件单元测试
- 清理服务新增 `FOLDER_PURGED_EVENT_FILE_IDS_LIMIT` 阈值（1000），超限时降级为 `SyncRequired` 事件以避免超大 payload；`purge_all_in_scope_silent` 改为遇错即返回而非仅打印警告
- 文件类型检测优先按扩展名判断，修复 `.ts` 文件被误识别为视频类型的问题
- E2E 测试等待上传完成后再继续，修复对话框关闭检测逻辑，代码预览改用 `pre` 元素断言
