+++ src/services/auth_service/tokens.rs
@@ -1,29 +1,23 @@
 //! 认证服务子模块：`tokens`。
 
 use chrono::{Duration as ChronoDuration, Utc};
-use dashmap::{DashMap, mapref::entry::Entry};
 use jsonwebtoken::{DecodingKey, EncodingKey, Header, Validation, decode, encode};
 use sea_orm::{ActiveValue::Set, ConnectionTrait};
-use serde::Serialize;
-use std::sync::LazyLock;
 
 use crate::config::auth_runtime::RuntimeAuthPolicy;
-use crate::db::repository::{auth_session_repo, user_repo};
+use crate::db::repository::auth_session_repo;
 use crate::entities::{auth_session, user};
 use crate::errors::{AsterError, MapAsterErr, Result};
 use crate::runtime::PrimaryAppState;
-use crate::services::audit_service::{self, AuditContext};
 use crate::types::TokenType;
 use crate::utils::numbers::{i64_to_u64, u64_to_i64, u64_to_usize};
 
-use super::session::{
-    get_auth_snapshot, invalidate_auth_snapshot_cache, purge_all_auth_sessions_in_connection,
-};
+use super::session::get_auth_snapshot;
 use super::{AuthSnapshot, Claims};
 
-const REFRESH_REUSE_GRACE_SECS: i64 = 15;
+mod refresh;
 
-static REFRESH_ROTATION_LOCKS: LazyLock<DashMap<String, ()>> = LazyLock::new(DashMap::new);
+pub use refresh::refresh_tokens;
 
 #[derive(Debug)]
 struct IssuedTokens {
@@ -32,34 +26,6 @@ struct IssuedTokens {
     session_id: String,
     refresh_jti: String,
     refresh_expires_at: chrono::DateTime<Utc>,
-}
-
-#[derive(Debug)]
-enum RefreshRotationError {
-    Aster(AsterError),
-    StaleRefresh { user_id: i64, reused_jti: String },
-    ReuseDetected { user_id: i64, reused_jti: String },
-}
-
-#[derive(Serialize)]
-struct RefreshTokenReuseAuditDetails<'a> {
-    reused_jti: &'a str,
-}
-
-struct RefreshRotationLockGuard {
-    refresh_jti: String,
-}
-
-impl Drop for RefreshRotationLockGuard {
-    fn drop(&mut self) {
-        REFRESH_ROTATION_LOCKS.remove(&self.refresh_jti);
-    }
-}
-
-impl From<AsterError> for RefreshRotationError {
-    fn from(value: AsterError) -> Self {
-        Self::Aster(value)
-    }
 }
 
 fn ensure_token_type(claims: &Claims, expected: TokenType) -> Result<()> {
@@ -79,101 +45,6 @@ fn ensure_session_current(claims: &Claims, snapshot: AuthSnapshot) -> Result<()>
     }
 
     Ok(())
-}
-
-fn is_recent_refresh_rotation(session: &auth_session::Model, now: chrono::DateTime<Utc>) -> bool {
-    now.signed_duration_since(session.last_seen_at)
-        .num_seconds()
-        .abs()
-        <= REFRESH_REUSE_GRACE_SECS
-}
-
-fn refresh_client_matches(
-    session: &auth_session::Model,
-    ip_address: Option<&str>,
-    user_agent: Option<&str>,
-) -> bool {
-    let Some(stored_user_agent) = session.user_agent.as_deref() else {
-        return false;
-    };
-    let Some(current_user_agent) = user_agent else {
-        return false;
-    };
-    if stored_user_agent != current_user_agent {
-        return false;
-    }
-
-    session.ip_address.is_none()
-        || ip_address.is_none()
-        || session.ip_address.as_deref() == ip_address
-}
-
-fn is_stale_refresh_from_same_client(
-    session: &auth_session::Model,
-    user_id: i64,
-    now: chrono::DateTime<Utc>,
-    ip_address: Option<&str>,
-    user_agent: Option<&str>,
-) -> bool {
-    session.user_id == user_id
-        && session.revoked_at.is_none()
-        && is_recent_refresh_rotation(session, now)
-        && refresh_client_matches(session, ip_address, user_agent)
-}
-
-fn classify_refresh_reuse_session(
-    reused_auth_session: Option<&auth_session::Model>,
-    user_id: i64,
-    refresh_jti: &str,
-    now: chrono::DateTime<Utc>,
-    ip_address: Option<&str>,
-    user_agent: Option<&str>,
-) -> RefreshRotationError {
-    if reused_auth_session.is_some_and(|session| {
-        is_stale_refresh_from_same_client(session, user_id, now, ip_address, user_agent)
-    }) {
-        return RefreshRotationError::StaleRefresh {
-            user_id,
-            reused_jti: refresh_jti.to_string(),
-        };
-    }
-
-    RefreshRotationError::ReuseDetected {
-        user_id,
-        reused_jti: refresh_jti.to_string(),
-    }
-}
-
-fn try_acquire_refresh_rotation_lock(refresh_jti: &str) -> Option<RefreshRotationLockGuard> {
-    match REFRESH_ROTATION_LOCKS.entry(refresh_jti.to_string()) {
-        Entry::Occupied(_) => None,
-        Entry::Vacant(entry) => {
-            entry.insert(());
-            Some(RefreshRotationLockGuard {
-                refresh_jti: refresh_jti.to_string(),
-            })
-        }
-    }
-}
-
-async fn classify_refresh_rotation_lock_loss<C: ConnectionTrait>(
-    db: &C,
-    claims: &Claims,
-    refresh_jti: &str,
-    ip_address: Option<&str>,
-    user_agent: Option<&str>,
-) -> Result<RefreshRotationError> {
-    let now = Utc::now();
-    let reused_auth_session =
-        auth_session_repo::find_by_previous_refresh_jti(db, refresh_jti).await?;
-    Ok(classify_refresh_reuse_session(
-        reused_auth_session.as_ref(),
-        claims.user_id,
-        refresh_jti,
-        now,
-        ip_address,
-        user_agent,
-    ))
 }
 
 async fn authenticate_token(
@@ -318,259 +189,6 @@ pub async fn issue_tokens_for_user(
     issue_tokens_for_session(state, user.id, user.session_version, ip_address, user_agent).await
 }
 
-pub async fn refresh_tokens(
-    state: &PrimaryAppState,
-    refresh: &str,
-    ip_address: Option<&str>,
-    user_agent: Option<&str>,
-) -> Result<(String, String)> {
-    tracing::debug!("refreshing auth tokens");
-    let claims = verify_token(refresh, &state.config.auth.jwt_secret)?;
-    ensure_token_type(&claims, TokenType::Refresh)?;
-    let refresh_jti = claims
-        .jti
-        .clone()
-        .ok_or_else(|| AsterError::auth_token_invalid("refresh token missing jti"))?;
-    let Some(_rotation_lock_guard) = try_acquire_refresh_rotation_lock(&refresh_jti) else {
-        tracing::debug!(
-            user_id = claims.user_id,
-            reused_jti = refresh_jti,
-            "concurrent refresh token rotation lost same-token race"
-        );
-        let txn = crate::db::transaction::begin(&state.db).await?;
-        let lock_loss = classify_refresh_rotation_lock_loss(
-            &txn,
-            &claims,
-            &refresh_jti,
-            ip_address,
-            user_agent,
-        )
-        .await?;
-        return match lock_loss {
-            RefreshRotationError::StaleRefresh {
-                user_id,
-                reused_jti,
-            } => {
-                crate::db::transaction::rollback(txn).await?;
-                tracing::debug!(
-                    user_id,
-                    reused_jti,
-                    "stale refresh token reused within rotation grace window"
-                );
-                Err(AsterError::auth_token_invalid("stale refresh token"))
-            }
-            RefreshRotationError::ReuseDetected {
-                user_id,
-                reused_jti,
-            } => {
-                user_repo::bump_session_version(&txn, user_id).await?;
-                purge_all_auth_sessions_in_connection(&txn, user_id).await?;
-                crate::db::transaction::commit(txn).await?;
-                invalidate_auth_snapshot_cache(state, user_id).await;
-                tracing::warn!(
-                    user_id,
-                    reused_jti,
-                    "refresh token reuse detected after losing rotation lock; revoked all sessions"
-                );
-                audit_service::log(
-                    state,
-                    &AuditContext {
-                        user_id,
-                        ip_address: None,
-                        user_agent: None,
-                    },
-                    audit_service::AuditAction::UserRefreshTokenReuseDetected,
-                    crate::services::audit_service::AuditEntityType::User,
-                    Some(user_id),
-                    None,
-                    audit_service::details(RefreshTokenReuseAuditDetails {
-                        reused_jti: &reused_jti,
-                    }),
-                )
-                .await;
-                Err(AsterError::auth_token_invalid(
-                    "refresh token reuse detected",
-                ))
-            }
-            RefreshRotationError::Aster(error) => {
-                crate::db::transaction::rollback(txn).await?;
-                Err(error)
-            }
-        };
-    };
-
-    let txn = crate::db::transaction::begin(&state.db).await?;
-    let rotation = async {
-        let now = Utc::now();
-        let existing_auth_session = auth_session_repo::find_by_refresh_jti(&txn, &refresh_jti)
-            .await
-            .map_err(RefreshRotationError::from)?;
-        let reused_auth_session = if existing_auth_session.is_none() {
-            auth_session_repo::find_by_previous_refresh_jti(&txn, &refresh_jti)
-                .await
-                .map_err(RefreshRotationError::from)?
-        } else {
-            None
-        };
-        let user = user_repo::find_by_id(&txn, claims.user_id)
-            .await
-            .map_err(RefreshRotationError::from)?;
-        if !user.status.is_active() {
-            return Err(RefreshRotationError::from(AsterError::auth_forbidden(
-                "account is disabled",
-            )));
-        }
-        if claims.session_version != user.session_version {
-            return Err(RefreshRotationError::from(AsterError::auth_token_invalid(
-                "session revoked",
-            )));
-        }
-
-        let Some(existing_auth_session) = existing_auth_session else {
-            if reused_auth_session.as_ref().is_some_and(|session| {
-                session.user_id == claims.user_id && session.revoked_at.is_none()
-            }) {
-                let stale_refresh_from_same_client =
-                    reused_auth_session.as_ref().is_some_and(|session| {
-                        is_stale_refresh_from_same_client(
-                            session,
-                            claims.user_id,
-                            now,
-                            ip_address,
-                            user_agent,
-                        )
-                    });
-                if stale_refresh_from_same_client {
-                    return Err(RefreshRotationError::StaleRefresh {
-                        user_id: claims.user_id,
-                        reused_jti: refresh_jti,
-                    });
-                }
-                user_repo::bump_session_version(&txn, claims.user_id)
-                    .await
-                    .map_err(RefreshRotationError::from)?;
-                purge_all_auth_sessions_in_connection(&txn, claims.user_id)
-                    .await
-                    .map_err(RefreshRotationError::from)?;
-                return Err(RefreshRotationError::ReuseDetected {
-                    user_id: claims.user_id,
-                    reused_jti: refresh_jti,
-                });
-            }
-            return Err(RefreshRotationError::from(AsterError::auth_token_invalid(
-                "session revoked",
-            )));
-        };
-
-        if existing_auth_session.user_id != claims.user_id {
-            return Err(RefreshRotationError::from(AsterError::auth_token_invalid(
-                "invalid token",
-            )));
-        }
-        if existing_auth_session.revoked_at.is_some() {
-            return Err(RefreshRotationError::from(AsterError::auth_token_invalid(
-                "session revoked",
-            )));
-        }
-
-        let next_ip_address = ip_address.or(existing_auth_session.ip_address.as_deref());
-        let next_user_agent = user_agent.or(existing_auth_session.user_agent.as_deref());
-        let tokens = issue_tokens_for_session_id(
-            state,
-            user.id,
-            user.session_version,
-            Some(existing_auth_session.id.as_str()),
-        )
-        .map_err(RefreshRotationError::from)?;
-
-        if !auth_session_repo::rotate_refresh(
-            &txn,
-            &refresh_jti,
-            &tokens.refresh_jti,
-            tokens.refresh_expires_at,
-            next_ip_address,
-            next_user_agent,
-            now,
-        )
-        .await
-        .map_err(RefreshRotationError::from)?
-        {
-            // We already observed this jti as current inside this refresh attempt.
-            // A failed conditional rotate means another request won the same-token race,
-            // not that an already-stale token was presented later.
-            return Err(RefreshRotationError::StaleRefresh {
-                user_id: claims.user_id,
-                reused_jti: refresh_jti,
-            });
-        }
-
-        Ok::<_, RefreshRotationError>((
-            (tokens.access_token, tokens.refresh_token),
-            user.session_version,
-        ))
-    }
-    .await;
-
-    match rotation {
-        Ok((tokens, session_version)) => {
-            crate::db::transaction::commit(txn).await?;
-            tracing::debug!(
-                user_id = claims.user_id,
-                session_version,
-                "refreshed auth tokens"
-            );
-            Ok(tokens)
-        }
-        Err(RefreshRotationError::StaleRefresh {
-            user_id,
-            reused_jti,
-        }) => {
-            crate::db::transaction::rollback(txn).await?;
-            tracing::debug!(
-                user_id,
-                reused_jti,
-                "stale refresh token reused within rotation grace window"
-            );
-            Err(AsterError::auth_token_invalid("stale refresh token"))
-        }
-        Err(RefreshRotationError::ReuseDetected {
-            user_id,
-            reused_jti,
-        }) => {
-            crate::db::transaction::commit(txn).await?;
-            invalidate_auth_snapshot_cache(state, user_id).await;
-            tracing::warn!(
-                user_id,
-                reused_jti,
-                "refresh token reuse detected; revoked all sessions"
-            );
-            audit_service::log(
-                state,
-                &AuditContext {
-                    user_id,
-                    ip_address: None,
-                    user_agent: None,
-                },
-                audit_service::AuditAction::UserRefreshTokenReuseDetected,
-                crate::services::audit_service::AuditEntityType::User,
-                Some(user_id),
-                None,
-                audit_service::details(RefreshTokenReuseAuditDetails {
-                    reused_jti: &reused_jti,
-                }),
-            )
-            .await;
-            Err(AsterError::auth_token_invalid(
-                "refresh token reuse detected",
-            ))
-        }
-        Err(RefreshRotationError::Aster(error)) => {
-            crate::db::transaction::rollback(txn).await?;
-            Err(error)
-        }
-    }
-}
-
 pub async fn revoke_refresh_token(state: &PrimaryAppState, token: &str) -> Result<bool> {
     let claims = match verify_token(token, &state.config.auth.jwt_secret) {
         Ok(claims) => claims,
@@ -656,21 +274,6 @@ mod tests {
     use crate::types::TokenType;
 
     const SECRET: &str = "test_secret_32bytes_xxxxxxxxxxxxxxxxx";
-
-    fn make_auth_session(now: chrono::DateTime<Utc>) -> auth_session::Model {
-        auth_session::Model {
-            id: "session-id".to_string(),
-            user_id: 1,
-            current_refresh_jti: "next-jti".to_string(),
-            previous_refresh_jti: Some("refresh-jti".to_string()),
-            refresh_expires_at: now + ChronoDuration::days(1),
-            ip_address: Some("203.0.113.10".to_string()),
-            user_agent: Some("browser".to_string()),
-            created_at: now,
-            last_seen_at: now,
-            revoked_at: None,
-        }
-    }
 
     fn make_token(
         token_type: TokenType,
@@ -762,39 +365,5 @@ mod tests {
             role: crate::types::UserRole::User,
         };
         assert!(ensure_session_current(&claims, snapshot).is_ok());
-    }
-
-    #[test]
-    fn classify_refresh_reuse_session_treats_same_client_grace_as_stale() {
-        let now = Utc::now();
-        let session = make_auth_session(now);
-
-        let result = classify_refresh_reuse_session(
-            Some(&session),
-            1,
-            "refresh-jti",
-            now,
-            Some("203.0.113.10"),
-            Some("browser"),
-        );
-
-        assert!(matches!(result, RefreshRotationError::StaleRefresh { .. }));
-    }
-
-    #[test]
-    fn classify_refresh_reuse_session_treats_different_client_as_reuse() {
-        let now = Utc::now();
-        let session = make_auth_session(now);
-
-        let result = classify_refresh_reuse_session(
-            Some(&session),
-            1,
-            "refresh-jti",
-            now,
-            Some("203.0.113.11"),
-            Some("other-browser"),
-        );
-
-        assert!(matches!(result, RefreshRotationError::ReuseDetected { .. }));
     }
 }
